[PLAYER-5446] Move different util modules and constants file
