+++ sdk/react/src/constants.js
