+++ sdk/react/src/lib/collapsingBarUtils.js
